analytics/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
sqlalchemy
psycopg2-binary
pyjwt
python-multipart
python-dotenv